Fix circular dep + standardise requires
